Added the wx_TAB_TRAVERSAL identifier to the scrollable window widgets in the Prefs so that TAB will move from field to field.
Added code to the connection failure handler in NextWord to prompt with a "Try another connection?" question, Yes=Show connection dialog, No=Pterm stays open/disconnected, Cancel=Exit Pterm.  This gives us a failure mode that allows the user to try again, especially useful when unable to establish the initial connection.  Use case: if the ASCII port is the default connection and for some reason Cyberserv is not responding on that port, this allows the user to select CIU mode without resorting to registry hacking (or whatever serves that purpose under Apple/Unix).


git-svn-id: svn://akdesign.dyndns.org/dtcyber/trunk@740 d8f3d347-7192-4e35-a67b-6d701703d080