Add test option.
If a command line argument is used ending in .dat, that's taken to be the
name of a test data file.  A pterm trace file is a valid data file;
basically it looks for one word per line, in octal, at the start of the
line.  It ignores any extra text and any lines that don't start with a digit.

The -a switch requests ascii mode; otherwise classic mode is used to
interpret the test data.


git-svn-id: svn://akdesign.dyndns.org/dtcyber/trunk@1298 d8f3d347-7192-4e35-a67b-6d701703d080