New simpler synchronizer model.
This model has an input register cleared by the 10 MHz clock and set
from the input lines, RS flop style.

It then acts on the inputs on the rising edge of the 10 MHz clock, so
it takes action on the latched inputs (before they are cleared).  The
output is simpy derived from the action result ANDed with the clock
pulse; that makes the output pulses coincident with the clock which is
strictly speaking not per spec, but it works fine.  At some later
point that timing will get revised anyway...


git-svn-id: svn://akdesign.dyndns.org/dtcyber/trunk@1309 d8f3d347-7192-4e35-a67b-6d701703d080