Change the wire lists to contain what was in the originals.
Any changes needed are now in "correction files" which wlist.py knows
how to process.  The naming convention is that chassis<n>.corr files
fix errors in the original (missing wires, mismatches between
endpoints where the correct answer is clear, etc.).
chassis<n>model.corr files are changes made to make things work in the
VHDL modeling environment (and, eventually, in the VHDL synthesis).

In addition, change the clocking to remove the clock fanout trees and
instead feed the correct clock phase directly to each input that
expects a clock.  Utility setclock.py finds the clock tree and
generates the clock file.  The clock files are chassis<n>clocks.corr.

Also changed behaviorally modelled pieces -- dtsynchro and memory --
to use the correct clocks.

Note that the model needs a 20-phase clock, pretty much, both positive
and inverted.  The description makes it look like there is a four
phase clock but that's only approximately accurate even for the PP,
and for the CPU it's very clear there are more phases than that.  To
cover all cases, we have 20 phases corresponding to each 5 ns multiple
offset from the reference point, in true and negated forms.  There is
a new "sysclk" module that generates them all, and the wire lists
connect to a clock bus which has 40 wires, one for each phase.


git-svn-id: svn://akdesign.dyndns.org/dtcyber/trunk@2106 d8f3d347-7192-4e35-a67b-6d701703d080