Fixed this bug report:
Boot from floppy.  Enter shift-stop to terminate the program.  (Result is a blank window.)  Enter any key ... *crash*.  The blank window with the status bar still saying "booted from floppy" is a bit disconcerting.  It might make sense simply to close the window when the program exits.  If you prefer keeping it, changing the status line to say "program ended" or something like that would be helpful.

git-svn-id: svn://akdesign.dyndns.org/dtcyber/branches/split@2218 d8f3d347-7192-4e35-a67b-6d701703d080