Change the packaging of the Micrototur floppy image for Pterm help. For Mac and Windows, where there is an easy way to package resource files alongside the executable program, we carry the actual floppy image file.  That is then read into memory on first help invocation. For Linux, which does not have such a mechanism, we continue to keep the image data as compiled-in initialization data for the buffer.
The master for the help is now the floppy image file ptermhelp.mte;
the Linux .h file is generated at build time from that.  Having the
floppy image as part of the program makes it easier to test updates to
that code: all that is needed is to update that file in the executable
directory (resource directory) and restart pterm.


git-svn-id: svn://akdesign.dyndns.org/dtcyber/branches/split@2285 d8f3d347-7192-4e35-a67b-6d701703d080