V6.0.5:
NOTE: The Mac release of this version is the last one to support Mac
OS versions earlier than 10.10, and the last to include PowerPC
support.  Linux binary kits are no longer supplied; Linux users
should build Pterm from the source tar archive.

Bugfixes.

Moved "Enable control characters" profile setting to the Emulation
tab, which is a somewhat more obvious spot for it than the Title tab.
This does not apply to Mac, which does not have that setting.

Add scale 3.5 (if Retina) and scale 4 menu options.


git-svn-id: svn://akdesign.dyndns.org/dtcyber/trunk@2411 d8f3d347-7192-4e35-a67b-6d701703d080